chore(sdk): move typings around for clearer separation between clients and servers
