+++ packages/sdk/src/bot/index.ts
@@ -1,4 +1,3 @@
 export * from './definition'
 export * from './implementation'
-export * from './context'
 export * from './client'
