+++ packages/sdk/src/bot/types/index.ts
@@ -1,0 +1,2 @@
+export * from './generic'
+export * from './common'
