+++ packages/sdk/src/integration/index.ts
@@ -1,5 +1,5 @@
-export * from './context'
+export * from './server/context'
 export * from './definition'
 export * from './implementation'
 export * from './client'
-export * from './logger'
+export * from './server/logger'
